Allow warning on missing textures with newEffectManager(true)
Instead of erroring out when a texture is missing, you can create the
EffectManager with 'true' as the first argument to instead print a
warning on stderr.

Some efkproj files seem to have an ColorTexture element pointing to just
the directory of textures (eg: <ColorTexture>Texture/</ColorTexture>) or
pointing to another .efkproj file. Ignoring the texture in this case at
least lets the effect partially work.