Set magnification when creating Effect rather than using SetScale().
SetScale() doesn't always get inherited by all nodes in an effect, so
setting the scale in order to try to magnify may make effects look
wrong. Instead we can set the magnification in newEffect(), which
properly scales the whole effect.

I think it is still reasonable to call SetScale() after Play() in order
to make the y-axis match love's, but there could be a better way to do
that too.