Add LoveModelLoader.
Now models are loaded properly as well. We weren't actually creating a
model loader at all since it wasn't compiled in, so models couldn't be
loaded with or without the love filesystem. With the LoveModelLoader,
models should be loadable similar to textures.

We'll probably need a LoveMaterialLoader as well, but I haven't found
any effects with materials yet.