Only restore vertex attrib arrays for love.js.
The reason this only affects love.js is because of this code in
love/src/modules/graphics/OpenGL.cpp:

        if (gl.isCoreProfile())
        {
                glGenVertexArrays(1, &mainVAO);
                glBindVertexArray(mainVAO);
        }

love.js uses OpenGLES2, which is not a core profile, so there's no
vertex array bound. When Effekseer tries to restore the original OpenGL
state, the enabled vertex arrays then aren't switched back to what LOVE
is expecting.