Use a FileInferace instead of the various Loader classes.
Most of the Loader classes were identical, and it turns out we can just
use a custom FileInterface to provide access to the love filesystem for
those. This takes care of the EffectLoader, ModelLoader, and now the
MaterialLoader. The TextureLoader still uses its own custom class in
order to load the textures directly into love.