Allow overriding the y-axis inversion.
Love2d has different orthographic projections for canvases and drawing
on the screen. Effekseer needs to be able to match these in order to
draw accurately in both cases. Ideally we'd pull the projection matrix
straight from love2d (and then override the near/far clip planes), but I
don't see an easy way to do that.