Convert non-rgba8 textures into rgba8 before passing to Effekseer.
EffectManager was assuming all textures were love's rgba8 and setting
the Effekseer format to R8G8B8A8_UNORM. If a texture was actually
non-rgba8, this would cause graphical issues.

Fortunately we can convert a non-rgba8 texture to an rgba8 one with the
ImageData:paste() function, which is fairly straightforward.

This was discovered with some EVFX-Stoneforge which has some 16-bit
grayscale textures, which are loaded in love2d as rgba16.