Change Y-axis inversion in orthographic mode.
In order to get the Y-axis in Effekseer to match love2d's, we were
inverting it in the projection matrix in orthographic mode. However,
that causes billboard objects in the effect to be inverted. For
symmetric billboards it doesn't matter, but for things like hearts or
other symbols, they are obviously incorrect (upside-down). This is
readily apparent in the EVFX-Charmforge effects.

We still need the Y coordinate to match love2d's when playing an effect
(so playing an effect at (100, 100) overlaps a sprite drawn at (100,
100), for example). However, that can be handled by doing setting Y to
(windowHeight - Y) when playing or moving an effect.

The setInvert() function and projection matrix changes when it is set
are still needed to handle drawing Effekseer effects in love2d canvases.
To draw to a canvas, call setInvert(true), then draw to the canvas, then
setInvert(false).