Remove Shader patches.
The Shader patch was incorrectly reimported in a3ead05 when upgrading to
Effekseer 1.62c. The code was originally introduced in ee20ee5 for
Effekseer 1.53c, but was removed during the upgrade to 1.60c2 and moved
into EffectManager.cpp. Unfortunately I forgot that it was moved since
it was buried in the import commit and was re-introduced. It's redundant
with the EffectManager.cpp __EMSCRIPTEN__ blocks.